Conflicts:
	Documents/Guide/Frames/Installing.tex
	Documents/Guide/Frames/functions.tex
	Documents/Guide/Guide.pdf
	Documents/Guide/Guide.tex

+++ Model/t1UpgradeInc/TriangleMod.h
@@ -3,7 +3,7 @@
  
  * File:   t1UpgradeInc/TriangleMod.h
  *
- * Copyright (c) 2004-2015 by Stuart Ansell
+ * Copyright (c) 2004-2016 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -106,9 +106,9 @@ class TriangleMod : public constructSystem::ModBase
   // special for 5 sides!!
   virtual int getExitWindow(const size_t,std::vector<int>&) const;
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
 
   virtual void createAll(Simulation&,const attachSystem::FixedComp&,
